+++ apm-sniffer/apm-sdk-plugin/kafka-v1-plugin/src/main/java/org/apache/skywalking/apm/plugin/kafka/v1/ConsumerEnhanceRequiredInfo.java
@@ -16,17 +16,38 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.kafka.v11;
+package org.apache.skywalking.apm.plugin.kafka.v1;
 
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceConstructorInterceptor;
+import java.util.Collection;
+import java.util.List;
 import org.apache.skywalking.apm.util.StringUtil;
 
-public class ProducerConstructorInterceptor implements InstanceConstructorInterceptor {
+public class ConsumerEnhanceRequiredInfo {
+    private String brokerServers;
+    private String topics;
+    private long startTime;
 
-    @Override public void onConstruct(EnhancedInstance objInst, Object[] allArguments) {
-        ProducerConfig config = (ProducerConfig)allArguments[0];
-        objInst.setSkyWalkingDynamicField(StringUtil.join(';', config.getList("bootstrap.servers").toArray(new String[0])));
+    public void setBrokerServers(List<String> brokerServers) {
+        this.brokerServers =  StringUtil.join(';', brokerServers.toArray(new String[0]));
+    }
+
+    public void setTopics(Collection<String> topics) {
+        this.topics = StringUtil.join(';', topics.toArray(new String[0]));
+    }
+
+    public String getBrokerServers() {
+        return brokerServers;
+    }
+
+    public String getTopics() {
+        return topics;
+    }
+
+    public void setStartTime(long startTime) {
+        this.startTime = startTime;
+    }
+
+    public long getStartTime() {
+        return startTime;
     }
 }
